+++ Userland/Libraries/LibJS/Runtime/PromisePrototype.cpp
@@ -6,6 +6,7 @@
 
 #include <AK/Function.h>
 #include <LibJS/Interpreter.h>
+#include <LibJS/Runtime/AbstractOperations.h>
 #include <LibJS/Runtime/Error.h>
 #include <LibJS/Runtime/GlobalObject.h>
 #include <LibJS/Runtime/Promise.h>
