+++ Userland/Libraries/LibJS/Runtime/TypedArray.cpp
@@ -6,6 +6,7 @@
  */
 
 #include <AK/Checked.h>
+#include <LibJS/Runtime/AbstractOperations.h>
 #include <LibJS/Runtime/ArrayBuffer.h>
 #include <LibJS/Runtime/GlobalObject.h>
 #include <LibJS/Runtime/IteratorOperations.h>
