+++ Userland/Libraries/LibJS/Runtime/Value.h
@@ -17,6 +17,7 @@
 #include <AK/Types.h>
 #include <LibJS/Forward.h>
 #include <LibJS/Runtime/BigInt.h>
+#include <LibJS/Runtime/PrimitiveString.h>
 #include <math.h>
 
 // 2 ** 53 - 1
@@ -378,11 +379,6 @@ bool same_value(Value lhs, Value rhs);
 bool same_value_zero(Value lhs, Value rhs);
 bool same_value_non_numeric(Value lhs, Value rhs);
 TriState abstract_relation(GlobalObject&, bool left_first, Value lhs, Value rhs);
-Function* get_method(GlobalObject& global_object, Value, const PropertyName&);
-size_t length_of_array_like(GlobalObject&, const Object&);
-Function* species_constructor(GlobalObject&, const Object&, Function& default_constructor);
-Value require_object_coercible(GlobalObject&, Value);
-MarkedValueList create_list_from_array_like(GlobalObject&, Value, AK::Function<Result<void, ErrorType>(Value)> = {});
 
 struct ValueTraits : public Traits<Value> {
     static unsigned hash(Value value)
